Fixed DB warning when ObjC flag has not been set
