+++ Trace/Library/Database/Model/DBMetric.swift
@@ -1,5 +1,5 @@
 //
-//  DBMetric+CoreDataProperties.swift
+//  DBMetric.swift
 //  Trace
 //
 //  Created by Shams Ahmed on 23/07/2019.
@@ -10,7 +10,10 @@
 import Foundation
 import CoreData
 
-extension DBMetric {
+/// Internal use only
+@objc(DBMetric)
+@objcMembers
+internal final class DBMetric: NSManagedObject, Identifiable {
     
     // MARK: - Enum
     
@@ -31,11 +34,16 @@ extension DBMetric {
     /// Internal use only
     @NSManaged internal var date: Date
     
-    // MARK: - Fetch Request
+    // MARK: - Class - Fetch Request
 
     @nonobjc internal class func fetchRequest() -> NSFetchRequest<DBMetric> {
         return NSFetchRequest<DBMetric>(entityName: String(describing: DBMetric.self))
     }
-}
 
-extension DBMetric: Identifiable { }
+    // MARK: - Init
+    
+    /// Internal use only
+    override init(entity: NSEntityDescription, insertInto context: NSManagedObjectContext?) {
+        super.init(entity: entity, insertInto: context)
+    }
+}
